backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic==2.6.1

# Database and Vector Store
pymongo==4.6.1
motor==3.3.2

# LangChain and AI Models
langsmith>=0.1.0,<0.2.0
langchain-core>=0.1.0,<0.2.0
langchain>=0.1.0,<0.2.0
anthropic>=0.17.0,<1.0.0
langchain-anthropic>=0.1.1,<0.2.0

# Azure OpenAI Integration
openai==1.12.0
azure-identity==1.15.0

# Data Processing - Updated versions
numpy>=1.18.5,<1.24.0  # Fixed version to be compatible
pandas>=1.5.3,<2.0.0   # Using older pandas version compatible with numpy
scipy>=1.9.3,<2.0.0    # Compatible with numpy

# Utilities
requests==2.31.0
aiohttp==3.9.3
typing-extensions==4.9.0
bs4
PyPDF2
tiktoken==0.5.2
mangum
python-multipart==0.0.7

